+++ modules/automation-account/automationaccount.tf
@@ -6,6 +6,15 @@ resource "azurerm_automation_account" "automation_account" {
 
   sku_name = var.automation_account_sku
   tags     = merge(local.default_tags, var.extra_tags, var.automation_account_extra_tags)
+
+  dynamic "identity" {
+    for_each = var.automation_account_identity_type != null ? ["fake"] : []
+    content {
+      type         = var.automation_account_identity_type.type
+      identity_ids = var.automation_account_identity_type.identity_ids
+    }
+  }
+
 }
 
 # Link automation account to log analytics workspace