+++ modules/automation-account/outputs.tf
@@ -27,3 +27,8 @@ output "automation_account_dsc_server_endpoint" {
   description = "Azure Automation Account DSC Server Endpoint"
   value       = azurerm_automation_account.automation_account.dsc_server_endpoint
 }
+
+output "automation_account_identity" {
+  description = "Identity block with principal ID and tenant ID"
+  value       = azurerm_automation_account.automation_account.identity
+}