+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Added
+  * AZ-685: Added `identity` variable for automation module
+
 # v5.0.0 - 2022-03-11
 
 Breaking